[stinkytofu] Implement pure enum-based LogicalInstruction design
Replace class-hierarchy LogicalInstruction with a pure enum approach:
- Convert LogicalInstruction from inheritance to enum-based design
- Add specialData infrastructure for pure enum design
- Add LogicalInstructionData.hpp for instruction metadata
- Complete pure enum design with factory functions
- Clean up pure enum design: correct messages and add documentation
- Remove factory dispatch table from IntrinsicExpansionPass
- Add hardware source file dependencies for ISA generation
- Replace docs: comprehensive instruction addition guide
- Remove obsolete builder functions in GenLogicalIR
- Enhance LogicalInstructions with flag management and inline methods
- Add logical counting functionality for Python bindings
- Remove LogicalInstructionFactory and related references
- Rename and update intrinsic definition files
- Fix and enable IntrinsicExpansionPass with typed instructions
- Implement parseOpcode() for string-to-opcode conversion
- Rename HLIR namespace to logical for consistency
- Simplify ClampReluF32 intrinsic definition
- Enhance Python test integration in CMake
- Add Python test integration to CMake for StinkyTofu
- Move LogicalInstructionDefs.inc to src/ir/logical/
- Update .gitignore to refine generated file exclusions
- Enhance LogicalInstructions dump for Label handling
- Remove yaml-cpp dependency from gfxisa target
- Add LogicalToAsmMultiArchTest for multi-architecture support
- Disable comment alignment for StinkyAsmEmitter tests
- Remove TensorLoadToLds opcode and update LogicalInstruction structure
- Update developer guides to reflect C++ instruction cost system
- Remove YAML-based hardware configuration and related latency helper
- Add architecture-specific instruction costs for Gfx942, Gfx950, and Gfx1250
- Implement cost management for GPU architecture instructions
- Add architecture-specific register limits to GpuArch
- Implement MFMA, SMFMA, and MXMFMA mnemonic generation in ToStinkyAsmPass

+++ shared/stinkytofu/python_module/src/python_bindings.cpp
@@ -87,7 +87,7 @@ NB_MODULE(_stinkytofu, m)
     // However, users typically don't construct Register directly - they use helper functions.
     nb::class_<StinkyRegister>(m, "Register")
         .def(nb::init<>(), "Create a null register")
-        .def(nb::init<const std::string&, int, int>(),
+        .def(nb::init<const std::string&, uint32_t, uint16_t>(),
              nb::arg("type"),
              nb::arg("index"),
              nb::arg("count") = 1,
